requirements.txt
Flask==2.1.0
pymongo==4.2.0
python-dotenv==0.20.0